+++ src/elements/mod.rs
@@ -4,6 +4,7 @@ mod b_button;
 mod b_icon;
 mod b_notification;
 mod b_table;
+mod b_tag;
 mod b_title;
 
 pub use b_block::*;
@@ -12,4 +13,5 @@ pub use b_button::*;
 pub use b_icon::*;
 pub use b_notification::*;
 pub use b_table::*;
+pub use b_tag::*;
 pub use b_title::*;