+++ website/src/pages/elements/mod.rs
@@ -1,0 +1,3 @@
+mod tag_page;
+
+pub use tag_page::TagPage;