+++ website/src/pages/mod.rs
@@ -7,6 +7,8 @@ mod home_page;
 mod layout_page;
 mod not_found_page;
 
+pub mod elements;
+
 pub use columns_page::ColumnsPage;
 pub use components_page::ComponentsPage;
 pub use elements_page::ElementsPage;